refactor(theme): mueve modules/Theme a shared/preferences/theme
