+++ app/providers.tsx
@@ -5,8 +5,8 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { Toaster as SonnerToaster } from 'sonner'
 import { useEffect, useState, type ReactNode } from 'react'
 import SessionProvider from '@/auth/contexts/SessionProvider'
-import { ThemeProvider } from '@/modules/Theme/ThemeContext'
-import { darkTheme, lightTheme, type Theme } from '@/modules/Theme/themesList'
+import { ThemeProvider } from '@/shared/preferences/theme'
+import { darkTheme, lightTheme, type Theme } from '@/shared/preferences/theme'
 import { useLocalStorage } from '@/shared/hooks/useLocalStorage'
 import { useIsHydrated } from '@/shared/hooks/useIsHydrated'
 import { useUserPreffedLanguage } from '@/modules/LanguageToggle/useUserPreffedLanguage'
