+++ src/shared/preferences/theme/index.ts
@@ -1,0 +1,6 @@
+export type { Theme } from './model/themesList'
+export { themesList, lightTheme, darkTheme } from './model/themesList'
+export { ThemeContext, ThemeProvider } from './state/ThemeContext'
+export { useChangeTheme } from './hooks/useChangeTheme'
+export { ThemeSelection } from './ui/ThemeSelection'
+export { default as ThemePreview } from './ui/ThemePreview'
