+++ src/shared/preferences/theme/model/themesList.ts
