+++ src/shared/preferences/theme/ui/ThemePreview.tsx
