+++ src/views/Settings.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'react-i18next'
 import { USER_IS_IN } from '../shared/ui/organisms/userIsIn'
 import { ChangeBoardName } from '../modules/board/components/ChangeBoardName'
 import CreateReminder from '@/modules/TaskBoard/components/Reminder/CreateReminder'
-import { ThemeSelection } from '../modules/Theme/ThemeSelection'
+import { ThemeSelection } from '@/shared/preferences/theme'
 import { ReminderList } from '@/modules/TaskBoard/components/Reminder/ReminderList'
 import PageContainer from './PageContainer'
 import { ToggleTypeOfView } from '@/modules/TypeOfView/ToggleTypeOfView'
