+++ src/views/UserDashboardSettings.tsx
@@ -1,7 +1,7 @@
 import { USER_IS_IN } from '@/shared/ui/organisms/userIsIn'
 import PageContainer from './PageContainer'
 import { useTranslation } from 'react-i18next'
-import { ThemeSelection } from '@/modules/Theme/ThemeSelection'
+import { ThemeSelection } from '@/shared/preferences/theme'
 
 function UserDashboardSettings() {
 	const { t } = useTranslation()
